% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ssb_09292}
\alias{ssb_09292}
\title{SSB gender equality data}
\format{
\subsection{\code{ssb_09292}}{

A data frame with 199,716 rows and 9 columns:
\describe{
\item{...1}{A vector of numbers, 1:199716}
\item{region, Region}{Regions as names (region) and code (Region)}
\item{year, Tid}{Year}
\item{contents, ContentsCode}{What are the 'qualitative' content referred to by 'value'}
\item{value}{value of indicator (as specified by content, year and region)}
\item{NAstatus}{NAstatus}
...
}
}
}
\source{
\url{https://www.ssb.no/statbank/table/09292/}
}
\usage{
ssb_09292
}
\description{
A dataset downloaded from SSB's open access statbank.
}
\keyword{datasets}
