% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_clean_df.R
\name{ssb_clean_df}
\alias{ssb_clean_df}
\title{SSB import and create file by tabnumber and filetype}
\usage{
ssb_clean_df(df, include = NULL)
}
\arguments{
\item{df}{A data frame.}

\item{include}{Columnnumbers of columns to be included in the cleaned dataset (ready for conversion to microdata-script.)}
}
\value{
A file, either .csv or RDS, with the data from the table.
}
\description{
SSB import and create file by tabnumber and filetype
}
