% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_giga_pull.R
\name{ssb_giga_pull}
\alias{ssb_giga_pull}
\title{SSB Giga pull}
\usage{
ssb_giga_pull(tabnr, size = 10, sys_sleep = 2)
}
\arguments{
\item{tabnr}{A number (written as character string) referring to a SSB-table in SSBs statbank.}

\item{size}{An integer which controls the number of Regions to be imported at once.}

\item{sys_sleep}{An integer which controls how long the computer 'sleeps' before importing the next chunk of data.}
}
\value{
A data frame (tibble) containing the pulled data.
}
\description{
SSB Giga pull
}
