% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_import_n_create_file_by_tabnum.R
\name{ssb_import_n_create_file_by_tabnum}
\alias{ssb_import_n_create_file_by_tabnum}
\title{SSB import and create file by tabnumber and filetype}
\usage{
ssb_import_n_create_file_by_tabnum(
  tabnum,
  filetype = "csv",
  size = 10,
  sys_sleep = 2
)
}
\arguments{
\item{filetype}{A character string, currently either "csv" or "RDS", which selects the type of file to be saved.}

\item{size}{An integer which controls the number of Regions to be imported at once.}

\item{sys_sleep}{An integer which controls how long the computer 'sleeps' before importing the next chunk of data.}

\item{tablenr}{A number, written as character string, referring to a SSB-table in SSBs statbank.}
}
\value{
A file, either .csv or RDS, with the data from the table.
}
\description{
SSB import and create file by tabnumber and filetype
}
