% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_import_n_save_list_of_dfs_as_csv.R
\name{ssb_import_n_save_list_of_dfs_as_csv}
\alias{ssb_import_n_save_list_of_dfs_as_csv}
\title{SSB import and create file by tabnumber and filetype}
\usage{
ssb_import_n_save_list_of_dfs_as_csv(tabnum_list = NULL, file = NULL)
}
\arguments{
\item{tabnum_list}{A vector of type character, with numbers corresponding to SSB/statbank-tables. written as character string, referring to a SSB-table in SSBs statbank.}

\item{file}{A path to, and name of, txt-file containing table numbers (one per row) for SSB/statbank - tables.}
}
\value{
Several files (csv) with data from tables corresponding to tabnum_list.
}
\description{
SSB import and create file by tabnumber and filetype
}
