% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_makeurl.R
\name{ssb_makeurl}
\alias{ssb_makeurl}
\title{SSB makeurl}
\usage{
ssb_makeurl(tablenr)
}
\arguments{
\item{tablenr}{A number (written as character string) referring to a SSB-table in SSBs statbank.}
}
\value{
A char vector of length one with the corresponding webadress (api) of the table.
}
\description{
SSB makeurl
}
