% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_metatibby.R
\name{ssb_metatibby}
\alias{ssb_metatibby}
\title{SSB metatibby}
\usage{
ssb_metatibby(tablenr)
}
\arguments{
\item{tablenr}{A number (written as character string) referring to a SSB-table in SSBs statbank.}
}
\value{
A dataframe (tibble) sized 2 x k, with column 1 giving variable names, and column 2 giving a list column of additional metadata.
}
\description{
SSB metatibby
}
