% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssb_overview.R
\name{ssb_overview}
\alias{ssb_overview}
\title{SSB overview}
\usage{
ssb_overview(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
Prints an overview of what's inside a dataset. Variables with 100 or more unique values just return the number of unique values.
}
\description{
SSB overview
}
